Manual reweighting loop
Corrections calculates variables and pickles data to optimize reading time. Pickles too big to upload. Manual reweighting calculates the weights between the d0 and d0bar distributions for the first variable and then applies the weights to a second variable. Recalculates weights and applies them to the next variable and so on, iterating around the variables until the weights converge.